Extract Version Table into a new sub component
