+++ src/versioning_manager.ui/src/views/OrgOverview.vue
@@ -24,22 +24,7 @@
               </v-card-title>
               <v-card-text>
                 <h2>Current Version:</h2><span>1.0.0.0</span>
-                <v-data-table
-                  :headers="versionHeaders"
-                  :items="versionTable"
-                  hide-actions
-                  class="elevation-1"
-                >
-                  <template slot="items" slot-scope="props">
-                    <td>{{ formatVersion(props.item.Version) }}</td>
-                    <td>
-                      <v-tooltip bottom>
-                        <span slot="activator">{{ humanizeDate(props.item.CreatedDate) }}</span>
-                        <span>Actual Date: {{ formatDateToString(props.item.CreatedDate) }}</span>
-                      </v-tooltip>
-                    </td>
-                  </template>
-                </v-data-table>
+                <VersionHistoryTable></VersionHistoryTable>
               </v-card-text>
             </v-card>
           </v-flex>
@@ -58,27 +43,19 @@ import ProductService from '@/services/ProductService';
 import OrganizationService from '@/services/OrganizationService';
 import VersionService from '@/services/VersionService';
 import moment from 'moment'
+import VersionHistoryTable from '@/components/VersionHistoryTable.vue'
 
-@Component
+@Component({
+  components: {
+    VersionHistoryTable
+  }
+})
 export default class OrgOverview extends Vue {
   private org: Organization = {
     Id: -1,
     Name: '[Organization]',
   };
   private productList: Product[] = [];
-  //TODO: Specify type
-  private versionTable: any = [];
-  
-  //TODO: Specify type
-  private versionHeaders: any = [
-          {
-            text: 'Build Number',
-            align: 'left',
-            sortable: false,
-            value: 'version'
-          },
-          {text: 'Created Date', value: 'created_date'}
-        ];
 
   public loadOrgById(orgId: number){
     OrganizationService.getById(orgId)
@@ -92,39 +69,10 @@ export default class OrgOverview extends Vue {
       .catch((error) => console.log(error));
   }
 
-  public loadVersions(productId: number){
-    VersionService.getAll()
-      .then((data) => this.versionTable = data)
-      .catch((error) => console.log(error));
-  }
-
   public mounted() {
     const orgId: number = Number(this.$route.params.id);
     this.loadOrgById(orgId);
     this.loadProducts(orgId);
-
-    // TODO: Move this so that it is associated to the Products
-    this.loadVersions(orgId);
-  }
-
-  private formatVersion(version: Version) {
-    return (
-      version.Major +
-      '.' +
-      version.Minor +
-      '.' +
-      version.Build +
-      '.' +
-      version.Revision
-    );
-  }
-
-  private formatDateToString(date: Date) {
-    return moment(date).format('YYYY-MM-DD, hh:mm:ss');
-  }
-
-  private humanizeDate(date: Date) {
-    return moment(date).fromNow();
   }
 }
 </script>
